(Refact) move login, logout codes to auth/ directory
