+++ src/features/Member/auth/utils/axios.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosError } from 'axios';
-import { useAuthStore } from '@/features/Member/store/auth';
+import { useAuthStore } from '@/features/Member/auth/store/auth';
 import { publishErrorToast } from '@/lib/errorEvents';
-import { logout } from '@/features/Member/utils/logout';
+import { logout } from '@/features/Member/auth/utils/logout';
 
 export const instance = axios.create({
   baseURL: import.meta.env.VITE_API_BASE_URL, // .env
