+++ src/hooks/useFetchBookmarkStatus.ts
@@ -1,6 +1,6 @@
-import { useAuthStore } from '@/features/Member/store/auth';
+import { useAuthStore } from '@/features/Member/auth/store/auth';
 import { useCompanyStore } from '@/store/company';
-import { instance as axios } from '@/features/Member/utils/axios';
+import { instance as axios } from '@/features/Member/auth/utils/axios';
 import { useCallback } from 'react';
 
 export function useFetchBookmarkStatus() {
