+++ src/router/index.tsx
@@ -3,12 +3,12 @@ import { createBrowserRouter } from 'react-router-dom';
 import Main from '@/features/Map/Main';
 import CompanyDetail from '@/features/Company/CompanyDetail';
 import Mypage from '@/features/Member/Mypage';
-import Login from '@/features/Member/Login';
+import Login from '@/features/Member/auth/Login';
 import LoginRequired from '@/pages/LoginRequired';
 import ToBeContinued from '@/pages/ToBeContinued';
 import MainLayout from '@/layout/MainLayout';
 import NotFound from '@/pages/NotFound';
-import OAuthCallback from '@/features/Member/OAuthCallback';
+import OAuthCallback from '@/features/Member/auth/OAuthCallback';
 import Ranking from '@/features/Competition/Ranking';
 import Competition from '@/features/Competition/Competition';
